requirement.txt
Install outside virtualenv
--------------------------
python==3.12.2
virtualenv==20.25.1

Install in virtualenv
--------------------------
Flask==3.0.3
waitress==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2==2.9.9
GeoAlchemy2==0.15.2
Flask-Migrate==4.0.7
google-auth==2.34.0
requests==2.32.3
Flask-SocketIO==5.3.7

Extra Dependency
---------------
# For Linux only
gunicorn==23.0.0